FIX import bug at ObjectDeleteMessage deserialization
