+++ packages/syft/src/syft/core/node/common/node_service/object_delete/__init__.py
@@ -1,0 +1,2 @@
+# relative
+from .upload_service_messages import ObjectDeleteMessage  # noqa: F401
